+++ src/main/php/Executor.php
@@ -32,14 +32,6 @@ interface Executor
     public function out();
 
     /**
-     * returns the output stream to write data outputted by executed command to
-     *
-     * @return  \stubbles\streams\OutputStream
-     * @deprecated  since 3.0.0, use out() instead, will be removed with 4.0.0
-     */
-    public function getOutputStream();
-
-    /**
      * executes given command
      *
      * @param   string  $command